pyproject.toml
[tool.poetry]
name = "elinext"
version = "0.1.0"
description = ""
authors = ["Artem Kiselev <akiselev.prod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.2"
django-environ = "^0.8.1"
factory-boy = "^3.2.1"
celery = "^5.2.1"
cyrtranslit = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
